feat: Home assistant K3s w/ docker-compose
I can build a k3s cluster running home assistant that is available on my
local network at localhost:8089.  This is not a scalable path for a
single device service as I would need a single port and path based
routing, but this is a good start.

Running baremetal on a series of networked devices requires a service
such as metallb to allocate local network addresses to k8s services, but
that must be done on baremetal linux (i.e. a pi or dedicated linux
server).